Set up Framework for thresholding and tried a few options
Added a call to the cv::threshold function in order to do binary
thresholding. The current problem with this code is that the
thresholding is far too blocky and still only works conditionally
in the cases that worked before. Before Thursday try to use some
image smoothing which I think might help.